backend/requirements.txt
sqlalchemy
fastapi
uvicorn
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose 